// src/estrados-imss/dtos/estrado-imss-create.dto.ts
export class EstradosImssDto {
  cveNotificaciones: number;
  registroPatronal: string;
  razonSocial: string;
  desDomicilio: string;
  fecPublicacion: Date;
  fecInicioPublicacion: Date;
  fecFinPublicacion: Date;
  fecRetiroPublicacion: Date;
  fecRegistro: number;
  fecPublicacionCadena: string;
  cveDoctoAdjuntoAcuerdo: number;
  desNumOficioAcuerdo: string;
  desNombreArchivoAcuerdo: string;
  desTipoAdjuntoAcuerdo: string;
  cveDoctoAdjuntoDocumento: number;
  desNumOficioDocumento: string;
  desNombreArchivoDocumento: string;
  desTipoAdjuntoDocumento: string;
  desTipodocumento: string;
  desProceso: string;
  fechaRegistro: Date;
}